Don't fail if localStorage cannot be used.
Previously, if localStorage was not accessible for whatever reason, then
the initialization of the extension would fail. This would leave the
extension with a #request-list that had no width set and no ability to
resize it. This was okay so long as the URLs in the request list were
relatively short but for very long URLs the request list would take more
than the width of the frame that DevTools made available for it and
would thus be unusable.